working_new/requirements.txt
tensorflow==2.16.1
numpy
pillow
scikit-image
keras
